requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
minio==7.2.0
requests==2.31.0
pandas==2.1.3
schedule==1.2.1
python-dateutil==2.8.2
pydantic==2.5.2 